Move 'Control Page' into a 'Set Up Steps' folder.
