+++ src/app/set-up-steps/contorl-page/control-navbar/control-navbar.component.html
