+++ src/app/set-up-steps/contorl-page/scenarios/scenarios.component.css
